refactor(cli): remove cloud api helpers layer
